+++ Services/WorkflowManagement/IWorkflowManagementService.cs
@@ -29,7 +29,7 @@ namespace Ideku.Services.WorkflowManagement
         Task<IEnumerable<Event>> GetAllEventsAsync();
 
         // Workflow Selection
-        Task<Models.Entities.Workflow?> GetApplicableWorkflowAsync(int categoryId, string divisionId, string departmentId, decimal savingCost, int? eventId);
+        Task<Models.Entities.Workflow?> GetApplicableWorkflowAsync(int categoryId, string divisionId, string departmentId, long savingCost, int? eventId);
 
         // Stage & Approver Management
         Task<IEnumerable<User>> GetApproversForWorkflowStageAsync(int workflowId, int targetStage, string? targetDivisionId, string? targetDepartmentId);
